+++ lib/flutter_team_logger.dart
@@ -1,2 +1,3 @@
+export 'src/file_log_flush_observer.dart';
 export 'src/log_item.dart';
 export 'src/logs.dart';